+++ server/src/Routes/Customers.hs
@@ -16,6 +16,7 @@ import Control.Exception.Safe (throwM, Exception, try)
 import Control.Monad ((>=>), (<=<), when, void)
 import Control.Monad.IO.Class (liftIO)
 import Data.Aeson (ToJSON(..), FromJSON(..), (.=), (.:), (.:?), withObject, object)
+import Data.Digest.Pure.MD5 (md5)
 import Data.Int (Int64)
 import Data.List (partition)
 import Data.Maybe (fromMaybe)
@@ -46,6 +47,7 @@ import Server
 import Validation (Validation(..))
 
 import qualified Crypto.BCrypt as BCrypt
+import qualified Data.ByteString.Lazy as LBS
 import qualified Data.CAProvinceCodes as CACodes
 import qualified Data.ISO3166_CountryCodes as CountryCodes
 import qualified Data.StateCodes as StateCodes
@@ -277,11 +279,18 @@ loginRoute LoginParameters { lpEmail, lpPassword, lpCartToken, lpRemember } =
                     authorizationError
             Nothing ->
                 hashAnyways authorizationError
-    where hashAnyways returnValue = do
+    where
+        -- Hash the password without trying to validate it to prevent
+        -- mining of valid logins.
+        hashAnyways :: App a -> App a
+        hashAnyways returnValue = do
             hash <- liftIO . BCrypt.hashPasswordUsingPolicy BCrypt.slowerBcryptHashingPolicy
                 $ encodeUtf8 lpPassword
             const returnValue $! hash
-          validatePassword (Entity customerId customer) =
+        -- Try to validate the customer's password via BCrypt, falling
+        -- back to ZenCart's hashing scheme on failure.
+        validatePassword :: Entity Customer -> App Bool
+        validatePassword c@(Entity customerId customer) =
             let
                 hashedPassword =
                     encodeUtf8 $ customerEncryptedPassword customer
@@ -294,17 +303,42 @@ loginRoute LoginParameters { lpEmail, lpPassword, lpCartToken, lpRemember } =
             in
                 if isValid && usesPolicy then
                     return True
-                else if isValid then do
-                    maybeNewHash <- liftIO . BCrypt.hashPasswordUsingPolicy
-                        BCrypt.slowerBcryptHashingPolicy
-                        $ encodeUtf8 lpPassword
-                    newHash <- maybe
-                        (serverError $ err500 { errBody = "Misconfigured Hashing Policy" })
-                        (return . decodeUtf8) maybeNewHash
-                    runDB $ update customerId [CustomerEncryptedPassword =. newHash]
-                    return True
+                else if isValid then
+                    makeNewPasswordHash customerId >> return True
                 else
+                    validateZencartPassword c
+        -- Try to use Zencart's password hashing scheme to validate the
+        -- user's password. If it is valid, upgrade to the BCrypt hashing
+        -- scheme.
+        validateZencartPassword :: Entity Customer -> App Bool
+        validateZencartPassword (Entity customerId customer) =
+            case T.splitOn ":" (customerEncryptedPassword customer) of
+                [passwordHash, salt] ->
+                    if T.length salt == 2 then
+                        let isValid =
+                                T.pack (show . md5 . LBS.fromStrict . encodeUtf8 $ salt <> lpPassword)
+                                    == passwordHash
+                        in
+                            if isValid then
+                                makeNewPasswordHash customerId >> return True
+                            else
+                                return False
+                    else
+                        return False
+                _ ->
                     return False
+        -- Upgrade the Customer's password by hashing it using the current
+        -- BCrypt hashing policy.
+        makeNewPasswordHash :: CustomerId -> App ()
+        makeNewPasswordHash customerId = do
+            maybeNewHash <- liftIO . BCrypt.hashPasswordUsingPolicy
+                BCrypt.slowerBcryptHashingPolicy
+                $ encodeUtf8 lpPassword
+            newHash <- maybe
+                (serverError $ err500 { errBody = "Misconfigured Hashing Policy" })
+                (return . decodeUtf8) maybeNewHash
+            runDB $ update customerId [CustomerEncryptedPassword =. newHash]
+
 
 
 -- LOGOUT